pyproject.toml
[project]
name = "va-notetaking"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "ipykernel>=7.2.0",
    "langchain>=0.1.0",
    "langchain-anthropic>=1.3.3",
    "langchain-google-genai>=1.0.0",
    "langchain-openai>=0.1.0",
    "langgraph>=0.0.20",
    "openai>=2.20.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0.3",
    "requests>=2.31.0",
    "streamlit>=1.30.0",
    "streamlit-mic-recorder>=0.0.8",
    "structlog>=24.1.0",
]
